requirements.txt
python-dotenv
streamlit==1.32.2
pandas==1.5.3
plotly==5.9.0
langchain==0.2.3
langchain_experimental==0.0.60
langchain_google_genai
tabulate==0.9.0